Refactor routes into separate modules for posts and users
Moved post and user-related API endpoints from `main.py` to dedicated modules under `routes/` for better modularity and maintainability. Updated `main.py` to include the new routers and removed the endpoint definitions from it. Added utility for password hashing to avoid circular import.
