+++ backend/app/utils.py
@@ -1,0 +1,3 @@
+from passlib.context import CryptContext
+
+pwd_context = CryptContext(schemes=['bcrypt'], deprecated='auto')
